pyproject.toml
[tool.poetry]
name = "tagandtake-backend"
version = "0.1.0"
description = ""
authors = ["Daniel Visser <daniel_visser95@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
rest-framework-simplejwt = "^0.0.2"
django-cors-headers = "^4.3.1"
django-allauth = "^0.63.3"
premailer = "^3.10.0"
drf-spectacular = "^0.27.2"
boto3 = "^1.34.146"
django-storages = "^1.14.4"
python-dotenv = "^1.0.1"
stripe = "^10.4.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

